changed many config files to in total smaller model for testing, improved code and before removing joints from prediction

+++ factr/trainers/base.py
@@ -71,7 +71,7 @@ class BaseTrainer(ABC):
             return
         self.schedule.step()
 
-    def save_checkpoint(self, global_step, top_k=1):
+    def save_checkpoint(self, global_step, top_k=2, onlysave_latest=False):
         model = self.model
         model_weights = (
             model.module.state_dict() if isinstance(model, DDP) else model.state_dict()
@@ -84,15 +84,18 @@ class BaseTrainer(ABC):
             global_step=global_step,
         )
         
-        # Save current checkpoint
-        current_ckpt = f"ckpt_{global_step:06d}.ckpt"
-        torch.save(save_dict, current_ckpt)
-        
-        # Remove old checkpoints, keeping only the 2 most recent
-        ckpts = sorted([f for f in os.listdir('.') if f.startswith('ckpt_') and f.endswith('.ckpt')])
-        for old_ckpt in ckpts[:-top_k]:  # Keep last 2 checkpoints
-            os.remove(old_ckpt)
-        torch.save(save_dict, f"rollout/latest_ckpt.ckpt")
+        if not onlysave_latest:
+            # Save current checkpoint
+            current_ckpt = f"ckpt_{global_step:06d}.ckpt"
+            torch.save(save_dict, current_ckpt)
+            
+            # Remove old checkpoints, keeping only the 2 most recent
+            ckpts = sorted([f for f in os.listdir('.') if f.startswith('ckpt_') and f.endswith('.ckpt')])
+            for old_ckpt in ckpts[:-top_k]:  # Keep last x checkpoints
+                os.remove(old_ckpt)
+            torch.save(save_dict, f"rollout/latest_ckpt.ckpt")
+        else:
+            torch.save(save_dict, f"rollout/latest_ckpt.ckpt")
 
     def load_checkpoint(self, load_path):
         load_dict = torch.load(load_path, weights_only=False)
